ai_intern_assessment/challenge2_pitch_visualizer/requirements.txt
flask==3.0.3
anthropic==0.28.0
openai==1.30.1
nltk==3.8.1
requests==2.32.3
python-dotenv==1.0.1
